updated helper names to avoid confusion between killAvoid and kill functions
